Fix build warnings (part 3)

+++ src/libkeymap/diacr.c
@@ -17,12 +17,12 @@
 #include "contextP.h"
 #include "ksyms.h"
 
-int lk_diacr_exists(struct lk_ctx *ctx, unsigned int index)
+int lk_diacr_exists(struct lk_ctx *ctx, int index)
 {
 	return (lk_array_get_ptr(ctx->accent_table, index) != NULL);
 }
 
-int lk_get_diacr(struct lk_ctx *ctx, unsigned int index, struct lk_kbdiacr *dcr)
+int lk_get_diacr(struct lk_ctx *ctx, int index, struct lk_kbdiacr *dcr)
 {
 	struct lk_kbdiacr *ptr;
 
@@ -58,7 +58,7 @@ int lk_append_diacr(struct lk_ctx *ctx, struct lk_kbdiacr *dcr)
 	return 0;
 }
 
-int lk_add_diacr(struct lk_ctx *ctx, unsigned int index, struct lk_kbdiacr *dcr)
+int lk_add_diacr(struct lk_ctx *ctx, int index, struct lk_kbdiacr *dcr)
 {
 	struct lk_kbdiacr *ptr;
 
@@ -77,7 +77,7 @@ int lk_add_diacr(struct lk_ctx *ctx, unsigned int index, struct lk_kbdiacr *dcr)
 	return 0;
 }
 
-int lk_del_diacr(struct lk_ctx *ctx, unsigned int index)
+int lk_del_diacr(struct lk_ctx *ctx, int index)
 {
 	int rc;
 	rc = lk_array_unset(ctx->accent_table, index);
@@ -98,9 +98,9 @@ int lk_append_compose(struct lk_ctx *ctx, struct lk_kbdiacr *dcr)
 		direction = TO_UNICODE;
 #endif
 
-	dcr0.diacr  = convert_code(ctx, dcr->diacr, direction);
-	dcr0.base   = convert_code(ctx, dcr->base, direction);
-	dcr0.result = convert_code(ctx, dcr->result, direction);
+	dcr0.diacr  = (unsigned int) convert_code(ctx, (int) dcr->diacr, direction);
+	dcr0.base   = (unsigned int) convert_code(ctx, (int) dcr->base, direction);
+	dcr0.result = (unsigned int) convert_code(ctx, (int) dcr->result, direction);
 
 	return lk_append_diacr(ctx, &dcr0);
 }
